requirements.txt
aiohttp
cachetools
discord.py
nltk
numpy
Pillow
PoE.py
